+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.1.0] - UNRELEASED
+
+### Added
+
+- New option `responseFactory` to configure a custom factory implementing `Psr\Http\Message\ResponseFactoryInterface`
+- New option `streamFactory` to configure a custom factory implementing `Psr\Http\Message\StreamFactoryInterface`
+
+### Changed
+
+- Upgrade dev dependencies
+
 ## [1.0.0] - 2018-01-25
 
 ### Added